# backend-go/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateClaimRequest:
    properties:
      customer_id:
        type: string
      description:
        maxLength: 1000
        minLength: 10
        type: string
      vehicle_id:
        type: string
    required:
    - customer_id
    - description
    - vehicle_id
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
    type: object
  dto.UpdateClaimRequest:
    properties:
      description:
        maxLength: 1000
        minLength: 10
        type: string
    required:
    - description
    type: object
...
  entity.Claim:
    properties:
      approved_by:
        type: string
      created_at:
        type: string
      customer_id:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        type: string
      total_cost:
        type: number
      updated_at:
        type: string
      vehicle_id:
        type: string
    type: object
...
  entity.ClaimHistory:
    properties:
      changed_at:
        type: string
      changed_by:
        type: string
      claim_id:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ev-warranty.com
    name: API Support
  description: API for EV Warranty Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EV Warranty API
  version: "1.0"
paths:
...
  /claims:
    get:
      consumes:
      - application/json
      description: Retrieve all warranty claims
      produces:
      - application/json
      responses:
        "200":
          description: Claims retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Claim'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all claims
      tags:
      - claims
    post:
      consumes:
      - application/json
      description: Create a new warranty claim (SC Technician/Staff only)
      parameters:
      - description: Claim creation data
        in: body
        name: createClaimRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClaimRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Claim created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Claim'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new claim
      tags:
      - claims
  /claims/{id}:
    delete:
      consumes:
      - application/json
      description: 'Delete a claim by ID (SC Staff: hard delete, EVM Staff: soft delete)'
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a claim
      tags:
      - claims
    get:
      consumes:
      - application/json
      description: Retrieve a specific claim by its ID
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Claim retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Claim'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get claim by ID
      tags:
      - claims
    put:
      consumes:
      - application/json
      description: Update an existing claim by ID (SC Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Claim update data
        in: body
        name: updateClaimRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateClaimRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Claim updated successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Update a claim
      tags:
      - claims
...
  /claims/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a claim and update its status (SC Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim cancelled successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Cancel a claim
      tags:
      - claims
  /claims/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark a claim as completed (EVM Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim completed successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Complete a claim
      tags:
      - claims
  /claims/{id}/history:
    get:
      consumes:
      - application/json
      description: Retrieve the history of status changes for a specific claim
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Claim history retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.ClaimHistory'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get claim history
      tags:
      - claims
...
  /claims/{id}/review:
    post:
      consumes:
      - application/json
      description: Review and approve/reject a submitted claim (EVM Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim reviewed successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Review a claim
      tags:
      - claims
  /claims/{id}/submit:
    post:
      consumes:
      - application/json
      description: Submit a claim to EVM for review (SC Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim submitted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Submit a claim for review
      tags:
      - claims
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"